dashboard configuration fix.
