+++ package.json
@@ -11,7 +11,8 @@
   },
   "scripts": {
     "build": "webpack --mode production",
-    "start": "webpack-dev-server --mode development",
+    "start-awfy": "webpack-dev-server --define process.env.DASHBOARD='\"awfy\"' --mode development",
+    "start-awsy": "webpack-dev-server --define process.env.DASHBOARD='\"awsy\"' --mode development",
     "test": "jest",
     "test:coverage": "jest --coverage",
     "lint": "eslint --cache --ext mjs,jsx,js --format codeframe \".*.js\" src test",
